refactoring the reading page code
